Change double to float

+++ src/main/java/org/kata/sp/domain/promotion/PromotionModel.java
@@ -14,6 +14,6 @@ import lombok.ToString;
 @AllArgsConstructor
 public class PromotionModel {
     private int quantity;
-    private double priceQuantity;
+    private float priceQuantity;
     private String productName;
 }
